Allow customers to record & upload video reviews (#143)
* sign_out :user instead of an object

So that we don't need to worry about if `User.last` actually exists.

Otherwise it will fail with

```
Could not find a valid mapping for nil
```

* UploadContextsController

We will use this to fetch necessary configs for product review video
uploads.

ReviewForm is used in many contexts and it's hard to make sure that all
ancestors have that available.

* Propagate video back to ReviewForm for upload

* Add webm MIME type

* No-op: move new review emails to the model

* No-op: extract ProductReview::UpdateService

This will allow easier coordination to update / destroy a review video.

* No-op: simplify `Purchase#post_review`

* No-op: use kwargs for `Purchase#post_review`

* ProductReview::UpdateService now updates videos

* Add disabled state to VideoReview

Will use this to disable the controls during an upload

* Use font-mono for timer

So it doesn't jump around

* temp for upload

* Auto set filetype for VideoFile

* Make VideoFile compatible with WithFileProperties

* Make sure evaporate react to prop changes

* Record in either MP4 or WebM depending on compatibility

* Record in 16:9

* temp

* tmp

* new VideoState type

* Show recorded video

* Rename to ReviewVideoRecorder*

This aligns with the other naming more consistently.

* Fix typing

* Fix ProductReviewPresenter refactor

* Fix initial reviewMode

* Rename ReviewVideoRecorder*

* Fix flickering

* Show camera error to end user

* Force remount to reacquire stream

This simplifies reacquiring the stream for the following situation:

1. User records a video
2. User deletes the recorded video
3. Now we're back in the preview state but the stream does not start
   automatically.

* Fix eslint

* Remove `preview` state

* Disable some actions during recording

* Defer rendering to avoid SSR errors

* Move UploadContextsController

Review form is also exposed on the product page, where we need the route
to be accessible on creator subdomains too.

* Include video on reviews page

* Simpler upload progress given the space constraint

* Only one alive record of each status

* Update ProductReviewPresenter spec

* Update ProductPresenter::ProductProps spec

* Update Purchase spec

* Add webm classification as video type

* Make s3_utility paths available outside of gumroad.com

We will allow file uploads (specifically video reviews in this case)
from creator subdomains, too.

* Add key for preview <video>

So that it doesn't re-use the live-stream `<video>` which has
`srcObject` set on it.

* Rubocop

* Explictly record at 720p30

* Update reviewVideoRecorderBusy

* No-op: renaming

* Add a 10-minute limit for now

* Fix cancel upload style